requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Dance
Flask-JWT-Extended
Flask-SocketIO
Flask-Mail
Flask-WTF
celery
redis
psycopg2-binary
python-dotenv
reportlab
gunicorn
pytest
pytest-flask
anthropic
